refactor: consolidate access list retrieval into shared.go
- Moved the getAccessList function from relay.go and gas_tank_relay.go to shared.go for improved code organization and reusability.
- Updated the tokenRelay function to utilize the new getAccessList implementation, enhancing clarity and reducing redundancy in RPC client connections.
