+++ packages/contracts-bedrock/test/supersim/shared.go
@@ -15,6 +15,7 @@ import (
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/ethclient"
+	"github.com/ethereum/go-ethereum/rpc"
 )
 
 // Identifier matches the ICrossL2Inbox.Identifier struct
@@ -155,4 +156,27 @@ func sendAndWaitForTransaction(client *ethclient.Client, chainID *big.Int, pk *e
 	}
 
 	return receipt, nil
+}
+
+func getAccessList(id Identifier, payload []byte) (*types.AccessList, error) {
+	// As pointed out, we should use an admin RPC client, similar to relay.go
+	// The relay.go script connects to port 8420 for the supersim admin rpc.
+	rpcClient, err := rpc.Dial("http://localhost:8420")
+	if err != nil {
+		return nil, fmt.Errorf("failed to connect to supersim admin RPC: %w", err)
+	}
+	defer rpcClient.Close()
+
+	req := GetAccessListForIdentifierRequest{
+		Identifier: id,
+		Payload:    "0x" + common.Bytes2Hex(payload),
+	}
+
+	var result GetAccessListResponse
+	err = rpcClient.CallContext(context.Background(), &result, "admin_getAccessListForIdentifier", req)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get access list via admin_getAccessListForIdentifier: %w", err)
+	}
+
+	return &result.AccessList, nil
 }
